[chore] reservation service 삭제
