--- Reservation/src/main/java/com/rainbowgon/reservation/ReservationApplication.java
@@ -1,13 +1,0 @@
-package com.rainbowgon.reservation;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ReservationApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ReservationApplication.class, args);
-    }
-
-}
