--- Reservation/src/main/java/com/rainbowgon/reservation/domain/reservation/entity/VerifiedStatus.java
@@ -1,5 +1,0 @@
-package com.rainbowgon.reservation.domain.reservation.entity;
-
-public enum VerifiedStatus {
-    VERIFIED, NOT_VERIFIED
-}
