--- Reservation/src/main/java/com/rainbowgon/reservation/global/entity/BaseEntity.java
@@ -1,25 +1,0 @@
-package com.rainbowgon.reservation.global.entity;
-
-import lombok.Getter;
-import org.hibernate.annotations.ColumnDefault;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public abstract class BaseEntity {
-
-    @CreatedDate
-    @ColumnDefault("CURRENT_TIMESTAMP")
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-    @Enumerated(EnumType.STRING)
-    @ColumnDefault("'VALID'")
-    @Column(columnDefinition = "VARCHAR(10)")
-    private ValidStatus isValid = ValidStatus.VALID;
-}
