--- Reservation/src/main/java/com/rainbowgon/reservation/global/entity/ValidStatus.java
@@ -1,5 +1,0 @@
-package com.rainbowgon.reservation.global.entity;
-
-public enum ValidStatus {
-    VALID, DELETED
-}
