--- Reservation/src/main/java/com/rainbowgon/reservation/global/error/errorCode/BaseErrorCode.java
@@ -1,8 +1,0 @@
-package com.rainbowgon.reservation.global.error.errorCode;
-
-import com.rainbowgon.reservation.global.error.dto.ErrorReason;
-
-public interface BaseErrorCode {
-
-    ErrorReason getErrorReason();
-}
