--- Reservation/src/test/java/com/rainbowgon/reservation/ReservationApplicationTests.java
@@ -1,13 +1,0 @@
-package com.rainbowgon.reservation;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ReservationApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
