+++ reservation-service/src/main/java/com/rainbowgon/reservationservice/ReservationServiceApplication.java
@@ -1,0 +1,13 @@
+package com.rainbowgon.reservationservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReservationServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ReservationServiceApplication.class, args);
+    }
+
+}
