+++ reservation-service/src/main/java/com/rainbowgon/reservationservice/domain/reservation/entity/VerifiedStatus.java
@@ -1,0 +1,5 @@
+package com.rainbowgon.reservationservice.domain.reservation.entity;
+
+public enum VerifiedStatus {
+    VERIFIED, NOT_VERIFIED
+}
