+++ reservation-service/src/main/java/com/rainbowgon/reservationservice/global/entity/ValidStatus.java
@@ -1,0 +1,5 @@
+package com.rainbowgon.reservationservice.global.entity;
+
+public enum ValidStatus {
+    VALID, DELETED
+}
