+++ reservation-service/src/main/java/com/rainbowgon/reservationservice/global/error/errorCode/BaseErrorCode.java
@@ -1,0 +1,8 @@
+package com.rainbowgon.reservationservice.global.error.errorCode;
+
+import com.rainbowgon.reservationservice.global.error.dto.ErrorReason;
+
+public interface BaseErrorCode {
+
+    ErrorReason getErrorReason();
+}
