+++ reservation-service/src/test/java/com/rainbowgon/reservationservice/ReservationServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.rainbowgon.reservationservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ReservationServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
